+++ Backend/dashboardApp/dashboardApp.http
@@ -1,0 +1,6 @@
+@dashboardApp_HostAddress = http://localhost:5053
+
+GET {{dashboardApp_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
