--- Backend/dawnDashboard/bin/Debug/net7.0/dawnDashboard.runtimeconfig.json
@@ -1,19 +1,0 @@
-{
-  "runtimeOptions": {
-    "tfm": "net7.0",
-    "frameworks": [
-      {
-        "name": "Microsoft.NETCore.App",
-        "version": "7.0.0"
-      },
-      {
-        "name": "Microsoft.AspNetCore.App",
-        "version": "7.0.0"
-      }
-    ],
-    "configProperties": {
-      "System.GC.Server": true,
-      "System.Runtime.Serialization.EnableUnsafeBinaryFormatterSerialization": false
-    }
-  }
-}
